+++ agent/sysinfo/kernel_windows.go
@@ -1,0 +1,14 @@
+package sysinfo
+
+import (
+	"github.com/lodastack/agent/agent/common"
+)
+
+func FsKernelMetrics() (L []*common.Metric) {
+	return nil
+}
+
+// exec `ps` to get all process states
+func PsMetrics() (L []*common.Metric) {
+	return nil
+}