+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugfixes
 
 - [#53](https://github.com/lodastack/agent/pull/53): Don't block if agent queue is full
+- [#54](https://github.com/lodastack/agent/pull/54): check agent post metric name
 
 ## v0.1.2 [2017-05-02]